Testes 2

+++ tests/test_qtd_bronze.sql
@@ -1,3 +1,3 @@
 select *
 from {{ source('bronze', 'abastecimento_vendas') }}
-where cast(qtd as bigint) <= 0
+where cast(qtd as float) <= 0
